refactor: share title-category table cross-crate
citum-migrate's SQI pruning must invert the engine's ref-type to
title-category defaulting exactly, but couldn't reach the engine's
pub(crate) tables and had re-implemented them separately. Move
TitleCategory and the three classification functions into
citum-schema-style, next to TitlesConfig, so both crates consume the
same table via the citum-schema facade re-export.
