+++ crates/citum-schema-style/src/options/mod.rs
@@ -15,6 +15,7 @@ pub mod multilingual;
 pub mod processing;
 pub mod scoped;
 pub mod substitute;
+pub mod title_class;
 
 pub use crate::presets::{MultilingualConfigEntry, MultilingualPreset};
 pub use bibliography::{
@@ -493,6 +494,10 @@ pub enum PageRangeFormat {
 
 pub mod titles;
 
+pub use title_class::{
+    TitleCategory, classified_ref_types, container_title_category, parent_serial_title_category,
+    title_category,
+};
 pub use titles::{TextCase, TitleRendering, TitlesConfig, TitlesConfigEntry};
 
 /// Structured link options.
